Extract single Flask view into own blueprint
This matches the global blueprint style and naming that we have in
our other repos [1][2].

[1]: https://github.com/alphagov/notifications-admin/blob/master/app/main/__init__.py
[2]: https://github.com/alphagov/document-download-frontend/blob/master/app/main/__init__.py
