+++ app/__init__.py
@@ -5,8 +5,6 @@ from notifications_utils import logging
 from notifications_utils.clients.statsd.statsd_client import StatsdClient
 
 from app.celery.celery import NotifyCelery
-from app.models.alerts import Alerts
-from app.render import get_rendered_pages
 
 notify_celery = NotifyCelery()
 statsd_client = StatsdClient()
@@ -22,21 +20,14 @@ def create_app():
     environment = os.getenv('NOTIFY_ENVIRONMENT', 'development')
     application.config.from_object(configs[environment])
 
+    from app.main import bp as main
+    application.register_blueprint(main)
+
     from app.commands import setup_commands
     setup_commands(application)
 
     statsd_client.init_app(application)
     logging.init_app(application, statsd_client)
     notify_celery.init_app(application)
-
-    @application.route('/<path:key>', methods=['GET'])
-    def show_page(key):
-        alerts = Alerts.load()
-        rendered_pages = get_rendered_pages(alerts)
-
-        if key in rendered_pages:
-            return rendered_pages[key]
-
-        return application.send_static_file(key)
 
     return application
